Added room_first, room_last and fixed _pattern
-Added room_first and room_last. They were declared, but not
implemented. So they trew undefined reference when used.
-Fixed draw_set_line_pattern having wrong definition, so it also trew
undefined reference erro when used.

+++ ENIGMAsystem/SHELL/Graphics_Systems/General/GSstdraw.h
@@ -32,7 +32,7 @@ unsigned draw_get_alpha_test_ref_value();
 void draw_set_alpha_test(bool enable);
 void draw_set_alpha_test_ref_value(unsigned val);
 void draw_enable_alphablend(bool enable);
-void draw_set_line_pattern(unsigned short pattern, int scale);
+void draw_set_line_pattern(int pattern, int scale);
 void draw_point(gs_scalar x, gs_scalar y);
 void draw_point_color(gs_scalar x, gs_scalar y, int color);
 void draw_line(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2);
@@ -62,20 +62,20 @@ void draw_roundrect_color_ext(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scala
 void draw_roundrect_precise(gs_scalar x1, gs_scalar y1,gs_scalar x2, gs_scalar y2, float rad, bool outline);
 void draw_roundrect_precise_color(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2, float rad, int col1, int col2, bool outline);
 
-inline void draw_roundrect(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2, bool outline) { 
-	draw_roundrect(x1, y1, x2, y2, 10, false); 
+inline void draw_roundrect(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2, bool outline) {
+	draw_roundrect(x1, y1, x2, y2, 10, false);
 }
 
-inline void draw_roundrect_color(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2, int col1, int col2, bool outline) { 
-	draw_roundrect_color(x1, y1, x2, y2, 10, col1, col2, false); 
+inline void draw_roundrect_color(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2, int col1, int col2, bool outline) {
+	draw_roundrect_color(x1, y1, x2, y2, 10, col1, col2, false);
 }
 
-inline void draw_roundrect_precise(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2, bool outline) { 
-	draw_roundrect_precise(x1, y1, x2, y2, 10, false); 
+inline void draw_roundrect_precise(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2, bool outline) {
+	draw_roundrect_precise(x1, y1, x2, y2, 10, false);
 }
 
-inline void draw_roundrect_precise_color(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2, int col1, int col2, bool outline) { 
-	draw_roundrect_precise_color(x1, y1, x2, y2, 10, col1, col2, false); 
+inline void draw_roundrect_precise_color(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2, int col1, int col2, bool outline) {
+	draw_roundrect_precise_color(x1, y1, x2, y2, 10, col1, col2, false);
 }
 
 void draw_healthbar(gs_scalar x1, gs_scalar y1, gs_scalar x2, gs_scalar y2, float amount, int backcol, int mincol, int maxcol, int direction, bool showback, bool showborder);
@@ -92,7 +92,7 @@ void draw_polygon_end(bool outline, bool allowHoles=true);
 
 }
 
-namespace enigma 
+namespace enigma
 {
 
 ///Simple container class for a Vertex in a Polygon.
